+++ docs/changelog.rst
@@ -48,7 +48,7 @@ In newer versions of rocksdb a bunch of options were moved or removed.
 * Moved ``Options.block_size_deviation`` to ``BlockBasedTableFactory``
 * Moved ``Options.block_restart_interval`` to ``BlockBasedTableFactory``
 * Moved ``Options.whole_key_filtering`` to ``BlockBasedTableFactory``
-
+* Removed ``Options.table_cache_remove_scan_count_limit``
 
 New:
 ^^^^